+++ batch.yaml
@@ -25,10 +25,13 @@ inscriptions:
         metus est et odio. Nullam venenatis, urna et molestie vestibulum, orci
         mi efficitur risus, eu malesuada diam lorem sed velit. Nam fermentum
         dolor et luctus euismod.
+    destination: bc1qw508d6qejxtdg4y5r3zarvary0c5xw7kv8f3t4
 
   - file: token.json
     metaprotocol: brc-20
+    destination: bc1pxwww0ct9ue7e8tdnlmug5m2tamfn7q06sahstg39ys4c9f3340qqxrdu9k
 
   - file: tulip.png
     metadata:
       author: Satoshi Nakamoto
+    destination: bc1pdqrcrxa8vx6gy75mfdfj84puhxffh4fq46h3gkp6jxdd0vjcsdyspfxcv6
